Fix code-block generation
Resetting is_code to False at the beginning of every blank line was
clearly broken when processings multi-line code; removed.
Code should also ignore all other style rules (since a comment in
Python can begin with a '#', but so can markdown headers), so
moved it to the top of the big if-elif.